requirements.txt
fastapi
uvicorn[standard]
httpx
google-api-python-client>=2.0
google-auth>=2.0
